candidate_search uniformation and comparison w/ hw
Examples 1 to 6 had a different filename and a slightly different algorithm. In this commit I have
uniformed all the example under the same algorithm and filename. Moreover, I've created a script
which compares the results of the hw cosntruction (fixed point integers) with the floating point
matlab contructor. The results for all the known examples (1 to 29) are identical.